+++ mockup/patterns/upload/pattern.js
@@ -7,7 +7,6 @@
  *    relativePath(string): again, to be used with baseUrl to create upload url (null)
  *    initialFolder(string): UID of initial folder related items widget should have selected (null)
  *    currentPath(string): Current path related items is starting with (null)
- *    clickable(boolean): If you can click on container to also upload (false)
  *    className(string): value for class attribute in the form element ('upload')
  *    paramName(string): value for name attribute in the file input element ('file')
  *    ajaxUpload(boolean): true or false for letting the widget upload the files via ajax. If false the form will act like a normal form. (true)
@@ -71,7 +70,6 @@ define([
       ajaxUpload: true,
 
       paramName: 'file',
-      clickable: true,
       addRemoveLinks: false,
       autoCleanResults: true,
       previewsContainer: '.previews',
@@ -98,7 +96,6 @@ define([
       // TODO: find a way to make this work in firefox (and IE)
       $(document).bind('paste', function(e){
         var oe = e.originalEvent;
-        var target = $(oe.target);
         var items = oe.clipboardData.items;
         if (items) {
           for (var i = 0; i < items.length; i++) {
@@ -111,7 +108,6 @@ define([
       });
       // values that will change current processing
       self.currentPath = self.options.currentPath;
-      self.numFiles = 0;
       self.currentFile = 0;
 
       template = _.template(template, {_t: _t});
@@ -153,8 +149,9 @@ define([
       $('button.browse', self.$el).click(function(e) {
         e.preventDefault();
         e.stopPropagation();
-        // we trigger the dropzone dialog!
-        self.dropzone.hiddenFileInput.click();
+        if(self.dropzone.files.length < self.options.maxFiles){
+          self.dropzone.hiddenFileInput.click();
+        }
       });
 
       var dzoneOptions = this.getDzoneOptions();
@@ -174,18 +171,16 @@ define([
         throw e;
       }
 
-      self.dropzone.on('addedfile', function(file) {
-        self.showControls();
+      self.dropzone.on('maxfilesreached', function(){
+        self.showHideControls();
+      });
+
+      self.dropzone.on('addedfile', function(/* file */) {
+        self.showHideControls();
       });
 
       self.dropzone.on('removedfile', function() {
-        if (self.dropzone.files.length < 1) {
-          self.hideControls();
-        } else {
-          // Clear the "you can not upload any more files" message
-          var file = self.dropzone.files[0];
-          $(".dz-error-message span", file.previewElement).html("");
-        }
+        self.showHideControls();
       });
 
       self.dropzone.on('success', function(e, response){
@@ -209,22 +204,22 @@ define([
       }
 
       self.dropzone.on('complete', function(file) {
-        if (file.status === Dropzone.SUCCESS && self.dropzone.files.length < 1) {
-          self.hideControls();
+        if (file.status === Dropzone.SUCCESS && self.dropzone.files.length === 1) {
+          self.showHideControls();
         }
       });
 
       self.dropzone.on('error', function(file, response, xmlhr) {
-        if (typeof xmlhr !== "undefined" && xmlhr.status !== 403){
+        if (typeof(xmlhr) !== 'undefined' && xmlhr.status !== 403){
           // If error other than 403, just print a generic message
-          $(".dz-error-message span", file.previewElement).html("The file transfer failed");
+          $('.dz-error-message span', file.previewElement).html(_('The file transfer failed'));
         }
       });
 
       self.dropzone.on('totaluploadprogress', function(pct) {
         // need to caclulate total pct here in reality since we're manually
         // processing each file one at a time.
-        pct = ((((self.currentFile - 1) * 100) + pct) / (self.numFiles * 100)) * 100;
+        pct = ((((self.currentFile - 1) * 100) + pct) / (self.dropzone.files.length * 100)) * 100;
         self.$progress.attr('aria-valuenow', pct).css('width', pct + '%');
       });
 
@@ -242,14 +237,42 @@ define([
       }
     },
 
-    showControls: function() {
+    showHideControls: function(){
+      /* we do this delayed because this can be called multiple times
+         AND we need to do this hide/show AFTER dropzone is done with
+         all it's own events. This is NASTY but the only way we can
+         enforce some numFiles with dropzone! */
       var self = this;
-      $('.controls', self.$el).fadeIn('slow');
+      if(self._showHideTimeout){
+        clearTimeout(self._showHideTimeout);
+      }
+      self._showHideTimeout = setTimeout(function(){
+        self._showHideControls();
+      }, 50);
     },
 
-    hideControls: function() {
+    _showHideControls: function(){
       var self = this;
-      $('.controls', self.$el).fadeOut('slow');
+      var $controls = $('.controls', self.$el);
+      var $browse = $('.browse-select', self.$el);
+      var $input = $('.dz-hidden-input');
+
+      if(self.options.maxFiles){
+        if(self.dropzone.files.length < self.options.maxFiles){
+          $browse.show();
+          $input.prop('disabled', false);
+        }else{
+          $browse.hide();
+          $input.prop('disabled', true);
+        }
+      }
+      if(self.dropzone.files.length > 0){
+        $controls.fadeIn('slow');
+        var file = self.dropzone.files[0];
+        $('.dz-error-message span', file.previewElement).html('');
+      }else{
+        $controls.fadeOut('slow');
+      }
     },
 
     pathJoin: function() {
@@ -295,18 +318,13 @@ define([
     getDzoneOptions: function() {
       var self = this;
 
-      // clickable option
-      if (typeof(self.options.clickable) === 'string') {
-        if (self.options.clickable === 'true') {
-          self.options.clickable = true;
-        } else {
-          self.options.clickable = false;
-        }
-      }
+      // This pattern REQUIRE dropzone to be clickable
+      self.options.clickable = true;
 
       var options = $.extend({}, self.options);
       options.url = self.getUrl();
-      // XXX force to only upload one at a time,
+
+      // XXX force to only upload one to the server at a time,
       // right now we don't support multiple for backends
       options.uploadMultiple = false;
 
@@ -349,7 +367,6 @@ define([
           fileaddedClassName = self.options.fileaddedClassName,
           finished = options.finished;
 
-      self.numFiles = self.dropzone.files.length;
       self.currentFile = 0;
 
       function process() {